backend/requirements.txt
flask==2.2.5
flask-cors==4.0.0
pillow==10.1.0
numpy==1.24.3
opencv-python-headless==4.8.1.78 